invoice_api/requirements.txt
Flask
pytesseract
Pillow
spacy
